test(lib): add full test coverage for ccf
- seperate test file to mock aws-use.json data
